Replace script pubkey function into separated module * Add qrc20_script.rs * Move Qrc20Coin::transfer_output into qrc20.rs
