+++ mm2src/coins/qrc20/qrc20_tests.rs
@@ -925,106 +925,6 @@ fn test_generate_token_transfer_script_pubkey() {
 }
 
 #[test]
-fn test_encode_decode_contract_number() {
-    let numbers = vec![
-        // left is source number, right is expected encoded array
-        (0i64, vec![]),
-        (1, vec![1]),
-        (-1, vec![129]),
-        (40, vec![40]),
-        (-40, vec![168]),
-        (-127, vec![255]),
-        (127, vec![127]),
-        (-128, vec![128, 128]),
-        (128, vec![128, 0]),
-        (255, vec![255, 0]),
-        (-255, vec![255, 128]),
-        (256, vec![0, 1]),
-        (-256, vec![0, 129]),
-        (2500000, vec![160, 37, 38]),
-        (-2500000, vec![160, 37, 166]),
-        (i64::max_value(), vec![255, 255, 255, 255, 255, 255, 255, 127]),
-        (i64::min_value(), vec![0, 0, 0, 0, 0, 0, 0, 128, 128]),
-        (Opcode::OP_4 as i64, vec![84]),
-        (Opcode::OP_CALL as i64, vec![194, 0]),
-    ];
-
-    for (source, encoded) in numbers {
-        println!("{}", source);
-        let actual_encoded = encode_contract_number(source);
-        assert_eq!(actual_encoded, encoded);
-        let actual_decoded = unwrap!(decode_contract_number(&encoded));
-        assert_eq!(actual_decoded, source);
-    }
-}
-
-#[test]
-fn test_extract_gas_limit_gas_price() {
-    let script: Script = "5403a02526012844a9059cbb0000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde3000000000000000000000000000000000000000000000000000000003b9aca0014d362e096e873eb7907e205fadc6175c6fec7bc44c2".into();
-
-    let expected_gas_limit = 2_500_000;
-    let actual = unwrap!(extract_gas_from_script(&script, ExtractGasEnum::GasLimit));
-    assert_eq!(actual, expected_gas_limit);
-
-    let expected_gas_price = 40;
-    let actual = unwrap!(extract_gas_from_script(&script, ExtractGasEnum::GasPrice));
-    assert_eq!(actual, expected_gas_price);
-}
-
-#[test]
-fn test_extract_contract_call() {
-    let script: Script = "5403a02526012844a9059cbb0000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde3000000000000000000000000000000000000000000000000000000003b9aca0014d362e096e873eb7907e205fadc6175c6fec7bc44c2".into();
-
-    let to_addr: UtxoAddress = "qHmJ3KA6ZAjR9wGjpFASn4gtUSeFAqdZgs".into();
-    let to_addr = qrc20_addr_from_utxo_addr(to_addr);
-    let amount: U256 = 1000000000.into();
-    let function = ERC20_CONTRACT.function("transfer").unwrap();
-    let expected = function
-        .encode_input(&[Token::Address(to_addr), Token::Uint(amount)])
-        .unwrap();
-
-    let actual = unwrap!(extract_contract_call_from_script(&script));
-    assert_eq!(actual, expected);
-
-    // TX b11a262380657310abf01f8abe117da2c2adf788ab1fa0fa29da4ab505fc00c0
-    let tx = unwrap!(hex::decode("01000000029ba0865fc62aac1f5f1a4aac3c9f54ff3d74211030bf6eb41e870b30297bd3fc010000006a47304402201808cbc98036ea63d32e858f776c722897d3f4b670744594deba25b69128d0ba02207b3f86f0ab6b6fa0ff581dc7be33af034c6004f3537a2f96c4ddf3ed0130defc012103693bff1b39e8b5a306810023c29b95397eb395530b106b1820ea235fd81d9ce9ffffffff63574ffa1e8edd8af8b08f3c1d8e5f33170772c38631a50ddc29c16d74c762f6020000006b483045022100db6cf963f6be56f7c6004ede74d452b2c5932eb6b12094fe67fa4ff0b6f4406e02207acff9163588a0c58fa009f5e876f90817b4006c303095d9b6425aae2922a485012103693bff1b39e8b5a306810023c29b95397eb395530b106b1820ea235fd81d9ce9ffffffff040000000000000000625403a08601012844095ea7b3000000000000000000000000ba8b71f3544b93e2f681f996da519a98ace0107a000000000000000000000000000000000000000000000000000000000000000014d362e096e873eb7907e205fadc6175c6fec7bc44c20000000000000000625403a08601012844095ea7b3000000000000000000000000ba8b71f3544b93e2f681f996da519a98ace0107a0000000000000000000000000000000000000000000000000000000001312d0014d362e096e873eb7907e205fadc6175c6fec7bc44c20000000000000000e35403a0860101284cc49b415b2a65e285b98480fd7de696e9fb5bcb68ec9468dd906c683e38cabb8f39905675fa0000000000000000000000000000000000000000000000000000000001312d00000000000000000000000000d362e096e873eb7907e205fadc6175c6fec7bc440000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde30101010101010101010101010101010101010101000000000000000000000000000000000000000000000000000000000000000000000000000000005f6d80b814ba8b71f3544b93e2f681f996da519a98ace0107ac2e52fdd05000000001976a9149e032d4b0090a11dc40fe6c47601499a35d55fbb88ac82816d5f"));
-    let utxo_tx: UtxoTx = unwrap!(deserialize(tx.as_slice()));
-
-    // first output in "b11a262380657310abf01f8abe117da2c2adf788ab1fa0fa29da4ab505fc00c0"
-    // `approve` to 0 contract call
-    let expected = unwrap!(hex::decode("095ea7b3000000000000000000000000ba8b71f3544b93e2f681f996da519a98ace0107a0000000000000000000000000000000000000000000000000000000000000000"));
-    let script = utxo_tx.outputs[0].script_pubkey.clone().into();
-
-    let actual = unwrap!(extract_contract_call_from_script(&script));
-    assert_eq!(actual, expected);
-
-    // second output in "b11a262380657310abf01f8abe117da2c2adf788ab1fa0fa29da4ab505fc00c0"
-    // `approve` to 20000000 contract call
-    let expected = unwrap!(hex::decode("095ea7b3000000000000000000000000ba8b71f3544b93e2f681f996da519a98ace0107a0000000000000000000000000000000000000000000000000000000001312d00"));
-    let script = utxo_tx.outputs[1].script_pubkey.clone().into();
-
-    let actual = unwrap!(extract_contract_call_from_script(&script));
-    assert_eq!(actual, expected);
-
-    // third output in "b11a262380657310abf01f8abe117da2c2adf788ab1fa0fa29da4ab505fc00c0"
-    // `erc20Payment` 20000000 amount contract call
-    let expected = unwrap!(hex::decode("9b415b2a65e285b98480fd7de696e9fb5bcb68ec9468dd906c683e38cabb8f39905675fa0000000000000000000000000000000000000000000000000000000001312d00000000000000000000000000d362e096e873eb7907e205fadc6175c6fec7bc440000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde30101010101010101010101010101010101010101000000000000000000000000000000000000000000000000000000000000000000000000000000005f6d80b8"));
-    let script = utxo_tx.outputs[2].script_pubkey.clone().into();
-
-    let actual = unwrap!(extract_contract_call_from_script(&script));
-    assert_eq!(actual, expected);
-}
-
-#[test]
-fn extract_token_addr() {
-    let script: Script = "5403a02526012844a9059cbb0000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde3000000000000000000000000000000000000000000000000000000003b9aca0014d362e096e873eb7907e205fadc6175c6fec7bc44c2".into();
-    let expected = qrc20_addr_from_str("0xd362e096e873eb7907e205fadc6175c6fec7bc44").unwrap();
-
-    let actual = unwrap!(extract_token_addr_from_script(&script));
-    assert_eq!(actual, expected);
-}
-
-#[test]
 fn test_transfer_details_by_hash() {
     // priv_key of qXxsj5RtciAby9T7m98AgAATL4zTi4UwDG
     let priv_key = [
